extend PoolManger
Moved the LogPipeThread and the LogPipeHandler out of the module and into their own module.
Created an exception "SubprocessException" to hold and handle exceptions captured in the subprocesses.
Upgraded the put_async function to better manage the uploading of tasks
Taken control of the logic of the get function to better control the resilence from subprocess behaviour
Added functionality to clear the tasks within the sendQueue
Raised functions from the pool Manager into the Communication object
Added documentation for the Logging functions within PoolManager
Added Daemon property to the PoolManager
Converted the Base of the PoolManager over to processes
Extended the PoolManager to have the functions of a typical process
Added tests